% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_effect_sizes.R
\name{adjust_effect_sizes}
\alias{adjust_effect_sizes}
\title{Adjust effect sizes of DESeq2 results}
\usage{
adjust_effect_sizes(
  .data,
  DE_name,
  type = c("apeglm", "ashr", "normal"),
  COEF = NULL,
  DESIGN = NULL,
  ...
)
}
\arguments{
\item{.data}{A 'SummarizedExperiment'}

\item{DE_name}{the name of the DE results to shrink}

\item{type}{Character string, type of adjustment to perform}

\item{COEF}{a character string. Coefficient to adjust effect sizes for}

\item{DESIGN}{A formula or character string that can be converted to a formula for the DE calculations.}

\item{...}{passed on to DESeq2::lfcShrink()}
}
\value{
a SummarizedExperiment
}
\description{
Adjust effect sizes of DESeq2 results
}
\details{
The function allows for effect size adjustments after DESeq2 DE calculations. By default, specified .data and DE_name is enough to perform the adjustment. In case the DE analysis does not exist, it will be generated using the specified coefficient.
}
\examples{
T

}
